move S3Config into destination-s3; update deps
