+++ airbyte-integrations/connectors/destination-s3/build.gradle
@@ -13,7 +13,6 @@ dependencies {
     implementation project(':airbyte-config:models')
     implementation project(':airbyte-protocol:models')
     implementation project(':airbyte-integrations:bases:base-java')
-    implementation project(':airbyte-integrations:connectors:destination-jdbc')
     implementation files(project(':airbyte-integrations:bases:base-java').airbyteDocker.outputs)
 
     // csv
